+++ test/client_tests/client_transport_connector_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, the gRPC project authors. Please see the AUTHORS file
+// Copyright (c) 2020, the gRPC project authors. Please see the AUTHORS file
 // for details. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,15 +14,10 @@
 // limitations under the License.
 
 import 'dart:async';
-import 'dart:convert';
 
 import 'package:grpc/grpc.dart';
-import 'package:grpc/src/client/call.dart';
 import 'package:grpc/src/client/http2_connection.dart';
-import 'package:grpc/src/generated/google/rpc/status.pb.dart';
-import 'package:grpc/src/shared/status.dart';
 import 'package:http2/transport.dart';
-import 'package:protobuf/protobuf.dart';
 import 'package:test/test.dart';
 
 import '../src/client_utils.dart';
@@ -31,10 +26,10 @@ import '../src/utils.dart';
 void main() {
   const dummyValue = 0;
 
-  ClientHarness harness;
+  ClientTransportConnectorHarness harness;
 
   setUp(() {
-    harness = ClientHarness()..setUp();
+    harness = ClientTransportConnectorHarness()..setUp();
   });
 
   tearDown(() {
@@ -262,27 +257,6 @@ void main() {
     );
   });
 
-  test('Call throws decoded message', () async {
-    const customStatusCode = 17;
-    const customStatusMessage = 'エラー';
-    const encodedCustomStatusMessage = '%E3%82%A8%E3%83%A9%E3%83%BC';
-
-    void handleRequest(_) {
-      harness.toClient.add(HeadersStreamMessage([
-        Header.ascii('grpc-status', '$customStatusCode'),
-        Header.ascii('grpc-message', encodedCustomStatusMessage)
-      ], endStream: true));
-      harness.toClient.close();
-    }
-
-    await harness.runFailureTest(
-      clientCall: harness.client.unary(dummyValue),
-      expectedException:
-          GrpcError.custom(customStatusCode, customStatusMessage),
-      serverHandlers: [handleRequest],
-    );
-  });
-
   test('Call throws on response stream errors', () async {
     void handleRequest(_) {
       harness.toClient.addError('Test error');
@@ -442,70 +416,5 @@ void main() {
         'myauthority.com');
     expect(Http2ClientConnection('localhost', null, channelOptions).authority,
         'myauthority.com');
-  });
-
-  test(
-      'decodeStatusDetails should decode details into a List<GeneratedMessage> if base64 present',
-      () {
-    final decodedDetails = decodeStatusDetails(
-        'CAMSEGFtb3VudCB0b28gc21hbGwafgopdHlwZS5nb29nbGVhcGlzLmNvbS9nb29nbGUucnBjLkJhZFJlcXVlc3QSUQpPCgZhbW91bnQSRVRoZSByZXF1aXJlZCBjdXJyZW5jeSBjb252ZXJzaW9uIHdvdWxkIHJlc3VsdCBpbiBhIHplcm8gdmFsdWUgcGF5bWVudA');
-    expect(decodedDetails, isA<List<GeneratedMessage>>());
-    expect(decodedDetails.length, 1);
-  });
-
-  test(
-      'decodeStatusDetails should decode details into an empty list for an invalid base64 string',
-      () {
-    final decodedDetails = decodeStatusDetails('xxxxxxxxxxxxxxxxxxxxxx');
-    expect(decodedDetails, isA<List<GeneratedMessage>>());
-    expect(decodedDetails.length, 0);
-  });
-
-  test('decodeStatusDetails should handle a null input', () {
-    final decodedDetails = decodeStatusDetails(null);
-    expect(decodedDetails, isA<List<GeneratedMessage>>());
-    expect(decodedDetails.length, 0);
-  });
-
-  test('parseGeneratedMessage should parse out a valid Any type', () {
-    final status = Status.fromBuffer(base64Url.decode(
-        'CAMSEGFtb3VudCB0b28gc21hbGwafgopdHlwZS5nb29nbGVhcGlzLmNvbS9nb29nbGUucnBjLkJhZFJlcXVlc3QSUQpPCgZhbW91bnQSRVRoZSByZXF1aXJlZCBjdXJyZW5jeSBjb252ZXJzaW9uIHdvdWxkIHJlc3VsdCBpbiBhIHplcm8gdmFsdWUgcGF5bWVudA=='));
-    expect(status.details, isNotEmpty);
-
-    final detailItem = status.details.first;
-    final parsedResult = parseErrorDetailsFromAny(detailItem);
-    expect(parsedResult, isA<BadRequest>());
-
-    final castedResult = parsedResult as BadRequest;
-    expect(castedResult.fieldViolations, isNotEmpty);
-    expect(castedResult.fieldViolations.first.field_1, 'amount');
-    expect(castedResult.fieldViolations.first.description,
-        'The required currency conversion would result in a zero value payment');
-  });
-
-  test('Call should throw details embedded in the headers', () async {
-    final code = StatusCode.invalidArgument;
-    final message = 'amount too small';
-    final details =
-        'CAMSEGFtb3VudCB0b28gc21hbGwafgopdHlwZS5nb29nbGVhcGlzLmNvbS9nb29nbGUucnBjLkJhZFJlcXVlc3QSUQpPCgZhbW91bnQSRVRoZSByZXF1aXJlZCBjdXJyZW5jeSBjb252ZXJzaW9uIHdvdWxkIHJlc3VsdCBpbiBhIHplcm8gdmFsdWUgcGF5bWVudA';
-
-    void handleRequest(_) {
-      harness.toClient.add(HeadersStreamMessage([
-        Header.ascii('grpc-status', code.toString()),
-        Header.ascii('grpc-message', message),
-        Header.ascii('grpc-status-details-bin', details),
-      ], endStream: true));
-      harness.toClient.close();
-    }
-
-    await harness.runFailureTest(
-      clientCall: harness.client.unary(dummyValue),
-      expectedException: GrpcError.custom(
-        code,
-        message,
-        decodeStatusDetails(details),
-      ),
-      serverHandlers: [handleRequest],
-    );
   });
 }